fix job init
